+++ lib/api/presenters/channel_stats_pair_presenter.rb
@@ -1,0 +1,14 @@
+module Api
+  module Presenters
+    module ChannelStatsPairPresenter
+      include Roar::JSON::HAL
+      include Roar::Hypermedia
+      include Grape::Roar::Representer
+      include BasePresenter
+
+      property :user1
+      property :user2
+      property :count
+    end
+  end
+end