+++ lib/api/presenters/channel_stats_presenter.rb
@@ -19,6 +19,7 @@ module Api
       property :users_in_sups_count
       property :users_opted_in_count
       property :users_count
+      property :unique_pairs_count
       property :outcomes
     end
   end