+++ lib/models/channel.rb
@@ -319,7 +319,7 @@ class Channel
 
   def export!(root)
     super
-    stats.export!(root, 'stats', Api::Presenters::ChannelStatsPresenter)
+    stats.export!(root)
     super(root, 'users', Api::Presenters::UserPresenter, users)
     super(root, 'rounds', Api::Presenters::RoundPresenter, rounds)
     super(root, 'sups', Api::Presenters::SupPresenter, sups)