+++ lib/models/channel_stats_pair.rb
@@ -1,0 +1,7 @@
+class ChannelStatsPair
+  include Mongoid::Document
+
+  field :user1, type: String
+  field :user2, type: String
+  field :count, type: Integer
+end