+++ lib/models/team_stats.rb
@@ -3,7 +3,7 @@ class TeamStats
   include SlackSup::Models::Mixins::Pluralize
   include SlackSup::Models::Mixins::Export
 
-  attr_accessor :channels_count, :channels_enabled_count, :rounds_count, :sups_count, :users_in_sups_count, :users_opted_in_count, :users_count, :outcomes, :team
+  attr_accessor :channels_count, :channels_enabled_count, :rounds_count, :sups_count, :users_in_sups_count, :users_opted_in_count, :users_count, :pairs, :outcomes, :team
 
   def initialize(team)
     @team = team
@@ -25,6 +25,27 @@ class TeamStats
         [(row['_id']['outcome'] || 'unknown').to_sym, row['count']]
       end
     ]
+
+    # https://stackoverflow.com/questions/37456062/how-to-get-combinations-of-items-in-an-array-field-in-mongodb
+
+    pairs_pipeline = [
+      { '$match': { channel_id: { '$in' => team.channels.distinct(:_id) } } },
+      { '$unwind': '$user_ids' },
+      { '$lookup': { from: 'sups', localField: '_id', foreignField: '_id', as: 'users' } },
+      { '$unwind': '$users' },
+      { '$unwind': '$users.user_ids' },
+      { '$redact': { '$cond': { if: { '$cmp': ['$user_ids', '$users.user_ids'] }, then: '$$DESCEND', else: '$$PRUNE' } } },
+      { '$group': { _id: { k1: '$user_ids', k2: '$users.user_ids' }, users: { '$sum': 0.5 } } },
+      { '$sort': { _id: 1 } },
+      { '$project':
+        { _id: 1, users: 1,
+          a: { '$cond': { if: { '$eq': [{ '$cmp': ['$_id.k1', '$_id.k2'] }, 1] }, then: '$_id.k2', else: '$_id.k1' } },
+          b: { '$cond': { if: { '$eq': [{ '$cmp': ['$_id.k1', '$_id.k2'] }, -1] }, then: '$_id.k2', else: '$_id.k1' } } } },
+      { '$group': { _id: { k1: '$a', k2: '$b' }, users: { '$sum': '$users' } } },
+      { '$project': { _id: 0, user1: '$_id.k1', user2: '$_id.k2', count: { '$convert': { input: '$users', to: 'int' } } } }
+    ]
+
+    @pairs = Sup.collection.aggregate(pairs_pipeline)
   end
 
   def positive_outcomes_count
@@ -33,6 +54,10 @@ class TeamStats
 
   def reported_outcomes_count
     outcomes.values.sum - (outcomes[:unknown] || 0)
+  end
+
+  def unique_pairs_count
+    pairs.count
   end
 
   def to_s
@@ -49,6 +74,7 @@ class TeamStats
       messages << "Facilitated #{pluralize(sups_count, 'S\'Up')} " \
                   "in #{pluralize(rounds_count, 'round')} " \
                   "for #{pluralize(users_count, 'user')} " \
+                  "creating #{pluralize(unique_pairs_count, 'unique connections')} " \
                   "with #{positive_outcomes_count * 100 / sups_count}% positive outcomes " \
                   "from #{reported_outcomes_count * 100 / sups_count}% outcomes reported."
     end